增加了虚拟文件系统用于管理JS插件模块
